Improves constant_folding2, replaces constant_folding.
