+++ compiler-lib/src/optimization/constant_folding2.rs
@@ -1,9 +1,11 @@
+use super::{OptimizationResult, OptimizationResultCollector};
 use crate::{debugging, firm::Program};
 use libfirm_rs::{
     graph::{Graph, NodeData},
-    nodes::{try_as_value_node, NodeTrait},
+    nodes::NodeTrait,
     nodes_gen::{Node, ProjKind},
     tarval::{Tarval, TarvalKind},
+    value_nodes::try_as_value_node,
 };
 use priority_queue::PriorityQueue;
 use std::{collections::hash_map::HashMap, path::PathBuf};
@@ -28,7 +30,8 @@ impl PartialOrd for Priority {
     }
 }
 
-pub fn run(program: &Program<'_, '_>) {
+pub fn run(program: &Program<'_, '_>) -> OptimizationResult {
+    let mut collector = OptimizationResultCollector::new();
     for class in program.classes.values() {
         for method in class.borrow().methods.values() {
             if let Some(graph) = method.borrow().graph {
@@ -36,14 +39,11 @@ pub fn run(program: &Program<'_, '_>) {
                 log::debug!("Graph for Method: {:?}", method.borrow().entity.name());
                 let mut cf = ConstantFolding::new(graph);
                 cf.run();
-                cf.apply();
-                /*graph.dump_dot_data(
-                    &PathBuf::from(format!("./dot-out/{}.dot", graph.entity().name_string())),
-                    |n| NodeData::new(format!("{:?}", n)),
-                );*/
+                cf.apply(&mut collector);
             }
         }
     }
+    collector.result()
 }
 
 #[derive(Debug, Clone, PartialEq, Copy)]
@@ -71,7 +71,10 @@ impl ConstantFolding {
 
         graph.walk_topological(|node| {
             topo_order += 1;
-            // blocks immediately, as they are not considered in walk_topological
+            // blocks and jumps are handled immediately,
+            // as they are not considered in walk_topological (in case of blocks)
+            // or don't depend much on anything before
+
             node_topo_idx.insert(
                 *node,
                 match node {
@@ -150,9 +153,10 @@ impl ConstantFolding {
         }
     }
 
-    fn dump_graph(&self, cur_node: Option<Node>)  {
+    // only for debugging/development
+    #[allow(dead_code)]
+    fn dump_graph(&self, cur_node: Option<Node>) {
         if self.graph.entity().name_string() == "CF.M.foo" {
-            // only for debugging/development
             self.graph.dump_dot_data(
                 &PathBuf::from(format!(
                     "./dot-out/{}.dot",
@@ -163,19 +167,19 @@ impl ConstantFolding {
                         Some(data) => format!("r: {:?}, val: {:?}", data.reachable, data.value),
                         None => " ".to_string(),
                     };
-                    match n {
-                        Node::Phi(phi) => {
-                            let mut string = "".to_string();
-                            for (pred, block_pred) in phi.in_nodes().zip(phi.block().in_nodes())
-                            {
-                                string += &format!("\n{:?} => {:?}", block_pred, pred);
-                            }
-                            str += &string;
+                    if let Node::Phi(phi) = n {
+                        let mut string = "".to_string();
+                        for (pred, block_pred) in phi.in_nodes().zip(phi.block().in_nodes()) {
+                            string += &format!("\n{:?} => {:?}", block_pred, pred);
                         }
-                        _ => {}
+                        str += &string;
                     }
-                    let mut nd =
-                        NodeData::new(format!("{:?}; {:?}\n{}", n, self.node_topo_idx.get(&n), str));
+                    let mut nd = NodeData::new(format!(
+                        "{:?}; {:?}\n{}",
+                        n,
+                        self.node_topo_idx.get(&n),
+                        str
+                    ));
                     nd.filled(Some(n) == cur_node);
                     if self.queue.get(&n).is_some() {
                         nd.bold(true)
@@ -260,19 +264,21 @@ impl ConstantFolding {
         MyLattice { reachable, value }
     }
 
-    fn apply(&mut self) {
-        if self.graph.entity().name_string() != "CF.M.foo" { return; }
-        self.dump_graph(None);
+    fn apply(&mut self, collector: &mut OptimizationResultCollector) {
+        //if self.graph.entity().name_string() != "CF.M.foo" { return; }
+        //self.dump_graph(None);
         let mut values = self.values.iter().collect::<Vec<_>>();
         values.sort_by_key(|(l, _)| l.node_id());
 
-        //let mut dangling_nontarget_blocks = Vec::new();
+        let mut dangling_blocks = Vec::new();
 
         for (node, lattice) in values {
             if !lattice.value.is_constant() {
+                collector.push(OptimizationResult::Unchanged);
                 continue;
             }
             if node.is_const() {
+                collector.push(OptimizationResult::Unchanged);
                 continue;
             }
 
@@ -280,39 +286,46 @@ impl ConstantFolding {
                 let const_node = self.graph.new_const(lattice.value);
                 log::debug!("EXCHANGE NODE {:?} val={:?}", node, lattice.value);
                 Graph::exchange_value(value_node.as_ref(), &const_node);
-            }
-            else {
-                match (node, lattice.value.kind()) {
-                    (Node::Cond(cond), TarvalKind::Bool(val)) => {
+                collector.push(OptimizationResult::Changed);
+            } else if let (Node::Cond(cond), TarvalKind::Bool(val)) = (node, lattice.value.kind()) {
+                let (always_taken_path, target_block) = cond.out_proj_target_block(val).unwrap();
 
-                        let (always_taken_path, target_block) = cond.out_proj_target_block(val).unwrap();
+                let (dead_path, nontarget_block) = cond.out_proj_target_block(!val).unwrap();
 
-                        let (dead_path, nontarget_block) = cond.out_proj_target_block(!val).unwrap();
-    /*
-    TODO see unrecahable_code_elimination
-                        if nontarget_block.num_cfgpreds() <= 1 {
-                            log::debug!("Mark nontarget block {:?} as dangling", nontarget_block);
-                            dangling_nontarget_blocks.push(nontarget_block);
-                        }
-    */
-
-                        let jmp = cond.block().new_jmp();
-                        log::debug!("Replace {:?} with {:?} to {:?}", node, jmp, target_block);
-                        Graph::exchange(&always_taken_path, &jmp);
-                        self.graph.mark_as_bad(&dead_path);
-
-                        // We need this because if we have a while(true) loop, the code will be
-                        // unreachable (libfirm-edge wise) from the end block (because the end block is
-                        // never reached control-flow wise), but libfirm needs to find the
-                        // loop (and it starts searching from the end block)
-                        target_block.keep_alive();
-                    }
-                    _ => {},
+                if nontarget_block.num_cfgpreds() <= 1 {
+                    // If the unused_proj is the sole predecessor of its successor,
+                    // mark the successor, eliminate it.
+                    // One would think this happens automatically, but it doesn't:
+                    // The successor block (if(f?) it contains a Jmp), is kept alive
+                    // somehow, causing be_lower_for_target to fail
+                    log::debug!("Mark nontarget_block {:?} as dangling", nontarget_block);
+                    dangling_blocks.push(nontarget_block);
                 }
+
+                let jmp = cond.block().new_jmp();
+                log::debug!("Replace {:?} with {:?} to {:?}", node, jmp, target_block);
+                Graph::exchange(&always_taken_path, &jmp);
+                self.graph.mark_as_bad(&dead_path);
+
+                // We need this because if we have a while(true) loop, the code will be
+                // unreachable (libfirm-edge wise) from the end block (because the end
+                // block is never reached control-flow wise), but
+                // libfirm needs to find the loop (and it starts
+                // searching from the end
+                // block)
+                target_block.keep_alive();
+                collector.push(OptimizationResult::Changed);
+            }
+        }
+
+        for block in &dangling_blocks {
+            for (i, child) in block.out_nodes().enumerate() {
+                log::debug!("Mark block child #{} {:?} as bad", i, child);
+                self.graph.mark_as_bad(&child);
             }
         }
 
         self.graph.remove_bads();
-        self.dump_graph(None);
+        //self.dump_graph(None);
     }
 }
